XPOS and FEATS from mate parser using Tueba-D/Z
The mate parser model was trained on the first 80% of Tueba-D/Z as
converted by an updated version of the TuebaUdConverter that improved
the morphological tag extraction. Tokens containing hyphens were
split into multiple tokens with flat/compound relations depending on
whether the original token was tagged as NE/other. The accuracy on
Tueba-D/Z test data for XPOS is 98.3% and for FEATS as a unit is 91.7%.